requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
reportlab==4.4.3
google-api-python-client==2.166.0
google-auth==2.38.0
